requirements.txt
python-dotenv
aquarium-python-api